w3c-reports/12092014/mac-wk2/runned-tests/2dcontext/drawing-images-to-the-canvas/2d.drawImage.svg-expected.txt

2d.drawImage.svg
drawImage() of an SVG image
Actual output:
Expected output:


FAIL drawImage() of an SVG image assert_approx_equals: Green channel of the pixel at (50, 25) expected 255 +/- 2 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:48:25
http://localhost:8800/2dcontext/drawing-images-to-the-canvas/2d.drawImage.svg.html:23:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:62:19
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/common/canvas-tests.js:59:15)
